+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## 1.1.0 - unreleased
+## 1.1.1 - unreleased
+
+- Fixed a deadlock while shutting down the StatusHandler
+
+## 1.1.0 - 2019-04-02
 
 - Added LookbackHandler
 - Added filtering / handling of specific loggers to multiprocess handlers